Test: refactor docs test to use fixtures
