+++ packages/tools/e2e-tests/src/support/page-objects/docs/components/menu.component.ts
@@ -1,18 +1,37 @@
+import { extractMetadataFromMarkdown } from '@helpers/docs/markdown.helper';
 import type { Locator, Page } from '@playwright/test';
 import { expect } from '@playwright/test';
 import dirTree from 'directory-tree';
 import path from 'path';
-import { extractMetadataFromMarkdown } from './markdown.helper';
 
-export default class TreeHelper {
+export default class MenuComponent {
   private readonly _page: Page;
-  private _parent: Locator;
+  private _levelTwo: Locator;
+  private _levelOne: Locator;
 
   public constructor(page: Page) {
     this._page = page;
-    this._parent = this._page.locator(
-      `[data-cy="sidemenu-submenu"] [data-testid="menuItem-1"]`,
+    this._levelOne = this._page.locator(
+      '[data-cy="sidemenu-submenu"] [data-testid="menuItem-1"] > button',
     );
+    this._levelTwo = this._page.locator(
+      '[data-cy="sidemenu-submenu"] [data-testid="menuItem-2"] > button',
+    );
+  }
+
+  public async openCollapsedItems(): Promise<void> {
+    const collapsedFirstLevelElements = await this._levelOne.all();
+
+    for (const collapsedElement of collapsedFirstLevelElements) {
+      await collapsedElement.click();
+      await this._page.waitForTimeout(1000);
+      await expect(collapsedElement).toHaveAttribute('data-active', 'true');
+    }
+    const collapsedSecondLevelElements = await this._levelTwo.all();
+    for (const collapsedElement of collapsedSecondLevelElements) {
+      await collapsedElement.click();
+      await expect(collapsedElement).toHaveAttribute('data-active', 'true');
+    }
   }
 
   /**
@@ -22,12 +41,13 @@ export default class TreeHelper {
   public async validateTree(pageToCheck: string): Promise<void> {
     const baseDir = path.join(
       __dirname,
-      '..',
-      '..',
-      '..',
-      '..',
-      '..',
-      '..',
+      '..', // components
+      '..', // docs
+      '..', // page-objects
+      '..', // support
+      '..', // e2e-tests
+      '..', // tools
+      '..', // packages
       'apps',
       'docs',
       'src',
